requirements.txt
Django==3.1
django-mptt==0.11.0
mysqlclient==2.0.1
